chore: split sidebar drop planner files
