+++ Sources/ContentView.swift
@@ -10113,7 +10113,9 @@ struct VerticalTabsSidebar: View {
                     selectedTabIds: $selectedTabIds,
                     lastSidebarSelectionIndex: $lastSidebarSelectionIndex,
                     dropIndicator: $dropIndicator,
-                    dragAutoScrollController: dragAutoScrollController,
+                    updateAutoscroll: {
+                        dragAutoScrollController.updateFromDragLocation()
+                    },
                     targets: renderContext.tabs.compactMap { tab in
                         guard let anchor = anchors[tab.id] else { return nil }
                         return SidebarDropPlanner.WorkspaceDropTarget(
@@ -12484,185 +12486,6 @@ private struct SidebarEmptyArea: View {
     }
 }
 
-private struct SidebarBonsplitTabWorkspaceDropOverlay: NSViewRepresentable {
-    let tabManager: TabManager
-    @Binding var selectedTabIds: Set<UUID>
-    @Binding var lastSidebarSelectionIndex: Int?
-    @Binding var dropIndicator: SidebarDropIndicator?
-    let dragAutoScrollController: SidebarDragAutoScrollController
-    let targets: [SidebarDropPlanner.WorkspaceDropTarget]
-
-    func makeNSView(context: Context) -> SidebarBonsplitTabWorkspaceDropView {
-        SidebarBonsplitTabWorkspaceDropView()
-    }
-
-    func updateNSView(_ nsView: SidebarBonsplitTabWorkspaceDropView, context: Context) {
-        nsView.targets = targets
-        nsView.isValidTransfer = {
-            guard let transfer = BonsplitTabDragPayload.currentTransfer() else { return false }
-            return AppDelegate.shared?.canMoveBonsplitTabToNewWorkspace(tabId: transfer.tab.id) ?? false
-        }
-        nsView.updateAutoscroll = {
-            dragAutoScrollController.updateFromDragLocation()
-        }
-        nsView.setDropIndicator = { indicator in
-            dropIndicator = indicator
-        }
-        nsView.performExistingWorkspaceMove = { workspaceId in
-            guard let transfer = BonsplitTabDragPayload.currentTransfer(),
-                  let app = AppDelegate.shared else {
-                return false
-            }
-            if let source = app.locateBonsplitSurface(tabId: transfer.tab.id),
-               source.workspaceId == workspaceId {
-                syncSidebarSelection()
-                return true
-            }
-            guard app.moveBonsplitTab(
-                tabId: transfer.tab.id,
-                toWorkspace: workspaceId,
-                focus: true,
-                focusWindow: true
-            ) else {
-                return false
-            }
-            selectedTabIds = [workspaceId]
-            syncSidebarSelection(preferredSelectedTabId: workspaceId)
-            return true
-        }
-        nsView.performNewWorkspaceMove = { insertionIndex, _ in
-            guard let transfer = BonsplitTabDragPayload.currentTransfer(),
-                  let app = AppDelegate.shared,
-                  let result = app.moveBonsplitTabToNewWorkspace(
-                    tabId: transfer.tab.id,
-                    destinationManager: tabManager,
-                    focus: true,
-                    focusWindow: true,
-                    insertionIndexOverride: insertionIndex
-                  ) else {
-                return false
-            }
-
-            selectedTabIds = [result.destinationWorkspaceId]
-            syncSidebarSelection(preferredSelectedTabId: result.destinationWorkspaceId)
-            return true
-        }
-    }
-
-    private func syncSidebarSelection(preferredSelectedTabId: UUID? = nil) {
-        let selectedId = preferredSelectedTabId ?? tabManager.selectedTabId
-        if let selectedId {
-            lastSidebarSelectionIndex = tabManager.tabs.firstIndex { $0.id == selectedId }
-        } else {
-            lastSidebarSelectionIndex = nil
-        }
-    }
-}
-
-private final class SidebarBonsplitTabWorkspaceDropView: NSView {
-    private static let pasteboardType = NSPasteboard.PasteboardType(BonsplitTabDragPayload.typeIdentifier)
-
-    var targets: [SidebarDropPlanner.WorkspaceDropTarget] = []
-    var isValidTransfer: () -> Bool = { false }
-    var updateAutoscroll: () -> Void = {}
-    var setDropIndicator: (SidebarDropIndicator?) -> Void = { _ in }
-    var performExistingWorkspaceMove: (UUID) -> Bool = { _ in false }
-    var performNewWorkspaceMove: (Int, SidebarDropIndicator) -> Bool = { _, _ in false }
-
-    override var isFlipped: Bool { true }
-    override var acceptsFirstResponder: Bool { false }
-
-    override init(frame frameRect: NSRect) {
-        super.init(frame: frameRect)
-        registerForDraggedTypes([Self.pasteboardType])
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func hitTest(_ point: NSPoint) -> NSView? {
-        shouldCaptureHitTest() ? super.hitTest(point) : nil
-    }
-
-    override func draggingEntered(_ sender: any NSDraggingInfo) -> NSDragOperation {
-        updateDrag(sender)
-    }
-
-    override func draggingUpdated(_ sender: any NSDraggingInfo) -> NSDragOperation {
-        updateDrag(sender)
-    }
-
-    override func draggingExited(_ sender: (any NSDraggingInfo)?) {
-        setDropIndicator(nil)
-    }
-
-    override func prepareForDragOperation(_ sender: any NSDraggingInfo) -> Bool {
-        acceptsDrag(sender) && SidebarDropPlanner.workspaceAction(for: localPoint(sender), targets: targets) != nil
-    }
-
-    override func performDragOperation(_ sender: any NSDraggingInfo) -> Bool {
-        defer { setDropIndicator(nil) }
-        guard acceptsDrag(sender),
-              let action = SidebarDropPlanner.workspaceAction(for: localPoint(sender), targets: targets) else {
-            return false
-        }
-
-        let moved: Bool
-        switch action {
-        case .existingWorkspace(let workspaceId):
-            moved = performExistingWorkspaceMove(workspaceId)
-        case .newWorkspace(let insertionIndex, let indicator):
-            moved = performNewWorkspaceMove(insertionIndex, indicator)
-        }
-
-        return moved
-    }
-
-    override func concludeDragOperation(_ sender: (any NSDraggingInfo)?) {
-        setDropIndicator(nil)
-    }
-
-    private func updateDrag(_ sender: any NSDraggingInfo) -> NSDragOperation {
-        guard acceptsDrag(sender) else {
-            setDropIndicator(nil)
-            return []
-        }
-
-        updateAutoscroll()
-        let point = localPoint(sender)
-        let action = SidebarDropPlanner.workspaceAction(for: point, targets: targets)
-        switch action {
-        case .newWorkspace(_, let indicator):
-            setDropIndicator(indicator)
-        case .existingWorkspace, nil:
-            setDropIndicator(nil)
-        }
-
-        return action == nil ? [] : .move
-    }
-
-    private func acceptsDrag(_ sender: any NSDraggingInfo) -> Bool {
-        guard sender.draggingPasteboard.types?.contains(Self.pasteboardType) == true else { return false }
-        return isValidTransfer()
-    }
-
-    private func shouldCaptureHitTest() -> Bool {
-        guard BonsplitTabDragPayload.currentTransfer() != nil else { return false }
-        guard let eventType = NSApp.currentEvent?.type else { return true }
-        switch eventType {
-        case .leftMouseDragged, .rightMouseDragged, .otherMouseDragged, .cursorUpdate, .mouseMoved:
-            return true
-        default:
-            return false
-        }
-    }
-
-    private func localPoint(_ sender: any NSDraggingInfo) -> CGPoint {
-        convert(sender.draggingLocation, from: nil)
-    }
-}
-
 enum SidebarPathFormatter {
     static let homeDirectoryPath: String = FileManager.default.homeDirectoryForCurrentUser.path
 
@@ -14807,226 +14630,6 @@ private struct SidebarMetadataMarkdownBlockRow: View {
             markdown: block.markdown,
             options: .init(interpretedSyntax: .full)
         )
-    }
-}
-
-enum SidebarDropEdge: Equatable {
-    case top
-    case bottom
-}
-
-struct SidebarDropIndicator: Equatable {
-    let tabId: UUID?
-    let edge: SidebarDropEdge
-}
-
-enum SidebarDropPlanner {
-    static func indicator(
-        draggedTabId: UUID?,
-        targetTabId: UUID?,
-        tabIds: [UUID],
-        pinnedTabIds: Set<UUID>,
-        pointerY: CGFloat? = nil,
-        targetHeight: CGFloat? = nil
-    ) -> SidebarDropIndicator? {
-        guard tabIds.count > 1, let draggedTabId else { return nil }
-        guard let fromIndex = tabIds.firstIndex(of: draggedTabId) else { return nil }
-
-        let insertionPosition: Int
-        if let targetTabId {
-            guard let targetTabIndex = tabIds.firstIndex(of: targetTabId) else { return nil }
-            let edge: SidebarDropEdge
-            if let pointerY, let targetHeight {
-                edge = edgeForPointer(locationY: pointerY, targetHeight: targetHeight)
-            } else {
-                edge = preferredEdge(fromIndex: fromIndex, targetTabId: targetTabId, tabIds: tabIds)
-            }
-            insertionPosition = (edge == .bottom) ? targetTabIndex + 1 : targetTabIndex
-        } else {
-            insertionPosition = tabIds.count
-        }
-
-        let legalInsertionPosition = legalInsertionPosition(
-            draggedTabId: draggedTabId,
-            proposedInsertionPosition: insertionPosition,
-            tabIds: tabIds,
-            pinnedTabIds: pinnedTabIds
-        )
-        let legalTargetIndex = resolvedTargetIndex(
-            from: fromIndex,
-            insertionPosition: legalInsertionPosition,
-            totalCount: tabIds.count
-        )
-        guard legalTargetIndex != fromIndex else { return nil }
-        return indicatorForInsertionPosition(legalInsertionPosition, tabIds: tabIds)
-    }
-
-    static func targetIndex(
-        draggedTabId: UUID,
-        targetTabId: UUID?,
-        indicator: SidebarDropIndicator?,
-        tabIds: [UUID],
-        pinnedTabIds: Set<UUID>
-    ) -> Int? {
-        guard let fromIndex = tabIds.firstIndex(of: draggedTabId) else { return nil }
-
-        let insertionPosition: Int
-        if let indicator, let indicatorInsertion = insertionPositionForIndicator(indicator, tabIds: tabIds) {
-            insertionPosition = indicatorInsertion
-        } else if let targetTabId {
-            guard let targetTabIndex = tabIds.firstIndex(of: targetTabId) else { return nil }
-            let edge = (indicator?.tabId == targetTabId)
-                ? (indicator?.edge ?? preferredEdge(fromIndex: fromIndex, targetTabId: targetTabId, tabIds: tabIds))
-                : preferredEdge(fromIndex: fromIndex, targetTabId: targetTabId, tabIds: tabIds)
-            insertionPosition = (edge == .bottom) ? targetTabIndex + 1 : targetTabIndex
-        } else {
-            insertionPosition = tabIds.count
-        }
-
-        let legalInsertionPosition = legalInsertionPosition(
-            draggedTabId: draggedTabId,
-            proposedInsertionPosition: insertionPosition,
-            tabIds: tabIds,
-            pinnedTabIds: pinnedTabIds
-        )
-        return resolvedTargetIndex(from: fromIndex, insertionPosition: legalInsertionPosition, totalCount: tabIds.count)
-    }
-
-    struct WorkspaceDropTarget: Equatable {
-        let workspaceId: UUID
-        let isPinned: Bool
-        let frame: CGRect
-    }
-
-    enum WorkspaceDropAction: Equatable {
-        case newWorkspace(insertionIndex: Int, indicator: SidebarDropIndicator)
-        case existingWorkspace(UUID)
-    }
-
-    static func workspaceAction(
-        for point: CGPoint,
-        targets: [WorkspaceDropTarget]
-    ) -> WorkspaceDropAction? {
-        guard !targets.isEmpty else { return nil }
-        let orderedTargets = targets.sorted { $0.frame.minY < $1.frame.minY }
-        if let containingTarget = orderedTargets.first(where: { $0.frame.contains(point) }) {
-            return workspaceAction(for: point, in: containingTarget, orderedTargets: orderedTargets)
-        }
-
-        guard let beforeTarget = orderedTargets.first(where: { point.y < $0.frame.minY }) else {
-            return nil
-        }
-        let insertionIndex = legalNewWorkspaceInsertionIndex(
-            orderedTargets.firstIndex(of: beforeTarget) ?? 0,
-            orderedTargets: orderedTargets
-        )
-        return .newWorkspace(
-            insertionIndex: insertionIndex,
-            indicator: workspaceIndicator(forInsertionIndex: insertionIndex, orderedTargets: orderedTargets)
-        )
-    }
-
-    private static func workspaceAction(
-        for point: CGPoint,
-        in target: WorkspaceDropTarget,
-        orderedTargets: [WorkspaceDropTarget]
-    ) -> WorkspaceDropAction? {
-        guard let targetIndex = orderedTargets.firstIndex(of: target) else { return nil }
-        let edgeBand = min(max(target.frame.height * 0.25, 10), target.frame.height / 2)
-        if point.y <= target.frame.minY + edgeBand {
-            let insertionIndex = legalNewWorkspaceInsertionIndex(targetIndex, orderedTargets: orderedTargets)
-            return .newWorkspace(
-                insertionIndex: insertionIndex,
-                indicator: workspaceIndicator(forInsertionIndex: insertionIndex, orderedTargets: orderedTargets)
-            )
-        }
-        if point.y >= target.frame.maxY - edgeBand {
-            let insertionIndex = legalNewWorkspaceInsertionIndex(targetIndex + 1, orderedTargets: orderedTargets)
-            return .newWorkspace(
-                insertionIndex: insertionIndex,
-                indicator: workspaceIndicator(forInsertionIndex: insertionIndex, orderedTargets: orderedTargets)
-            )
-        }
-        return .existingWorkspace(target.workspaceId)
-    }
-
-    private static func legalNewWorkspaceInsertionIndex(
-        _ proposedInsertion: Int,
-        orderedTargets: [WorkspaceDropTarget]
-    ) -> Int {
-        let clamped = max(0, min(proposedInsertion, orderedTargets.count))
-        let pinnedCount = orderedTargets.reduce(into: 0) { count, target in
-            if target.isPinned {
-                count += 1
-            }
-        }
-        return max(clamped, pinnedCount)
-    }
-
-    private static func workspaceIndicator(
-        forInsertionIndex insertionIndex: Int,
-        orderedTargets: [WorkspaceDropTarget]
-    ) -> SidebarDropIndicator {
-        let clampedInsertion = max(0, min(insertionIndex, orderedTargets.count))
-        if clampedInsertion >= orderedTargets.count {
-            return SidebarDropIndicator(tabId: nil, edge: .bottom)
-        }
-        return SidebarDropIndicator(tabId: orderedTargets[clampedInsertion].workspaceId, edge: .top)
-    }
-
-    private static func indicatorForInsertionPosition(_ insertionPosition: Int, tabIds: [UUID]) -> SidebarDropIndicator {
-        let clampedInsertion = max(0, min(insertionPosition, tabIds.count))
-        if clampedInsertion >= tabIds.count {
-            return SidebarDropIndicator(tabId: nil, edge: .bottom)
-        }
-        return SidebarDropIndicator(tabId: tabIds[clampedInsertion], edge: .top)
-    }
-
-    private static func insertionPositionForIndicator(_ indicator: SidebarDropIndicator, tabIds: [UUID]) -> Int? {
-        if let tabId = indicator.tabId {
-            guard let targetTabIndex = tabIds.firstIndex(of: tabId) else { return nil }
-            return indicator.edge == .bottom ? targetTabIndex + 1 : targetTabIndex
-        }
-        return tabIds.count
-    }
-
-    private static func preferredEdge(fromIndex: Int, targetTabId: UUID, tabIds: [UUID]) -> SidebarDropEdge {
-        guard let targetIndex = tabIds.firstIndex(of: targetTabId) else { return .top }
-        return fromIndex < targetIndex ? .bottom : .top
-    }
-
-    private static func legalInsertionPosition(
-        draggedTabId: UUID,
-        proposedInsertionPosition: Int,
-        tabIds: [UUID],
-        pinnedTabIds: Set<UUID>
-    ) -> Int {
-        let clampedInsertion = max(0, min(proposedInsertionPosition, tabIds.count))
-        guard !pinnedTabIds.isEmpty else { return clampedInsertion }
-
-        let pinnedCount = tabIds.reduce(into: 0) { count, tabId in
-            if pinnedTabIds.contains(tabId) {
-                count += 1
-            }
-        }
-        guard pinnedCount > 0 else { return clampedInsertion }
-
-        if pinnedTabIds.contains(draggedTabId) {
-            return min(clampedInsertion, pinnedCount)
-        }
-        return max(clampedInsertion, pinnedCount)
-    }
-
-    static func edgeForPointer(locationY: CGFloat, targetHeight: CGFloat) -> SidebarDropEdge {
-        guard targetHeight > 0 else { return .top }
-        let clampedY = min(max(locationY, 0), targetHeight)
-        return clampedY < (targetHeight / 2) ? .top : .bottom
-    }
-
-    private static func resolvedTargetIndex(from sourceIndex: Int, insertionPosition: Int, totalCount: Int) -> Int {
-        let clampedInsertion = max(0, min(insertionPosition, totalCount))
-        let adjusted = clampedInsertion > sourceIndex ? clampedInsertion - 1 : clampedInsertion
-        return max(0, min(adjusted, max(0, totalCount - 1)))
     }
 }
