+++ cmuxTests/SidebarOrderingTests.swift
@@ -804,102 +804,6 @@ final class SidebarDropPlannerTests: XCTestCase {
         XCTAssertEqual(targetIndex, 2)
     }
 
-    func testWorkspaceDropCenterTargetsExistingWorkspace() {
-        let first = UUID()
-        let second = UUID()
-        let targets = workspaceDropTargets([first, second])
-
-        let action = SidebarDropPlanner.workspaceAction(
-            for: CGPoint(x: 12, y: 56),
-            targets: targets
-        )
-
-        XCTAssertEqual(action, .existingWorkspace(second))
-    }
-
-    func testWorkspaceDropTopEdgeCreatesWorkspaceBeforeTarget() {
-        let first = UUID()
-        let second = UUID()
-        let targets = workspaceDropTargets([first, second])
-
-        let action = SidebarDropPlanner.workspaceAction(
-            for: CGPoint(x: 12, y: 42),
-            targets: targets
-        )
-
-        XCTAssertEqual(
-            action,
-            .newWorkspace(
-                insertionIndex: 1,
-                indicator: SidebarDropIndicator(tabId: second, edge: .top)
-            )
-        )
-    }
-
-    func testWorkspaceDropGapCreatesWorkspaceBeforeNextTarget() {
-        let first = UUID()
-        let second = UUID()
-        let targets = workspaceDropTargets([first, second])
-
-        let action = SidebarDropPlanner.workspaceAction(
-            for: CGPoint(x: 12, y: 36),
-            targets: targets
-        )
-
-        XCTAssertEqual(
-            action,
-            .newWorkspace(
-                insertionIndex: 1,
-                indicator: SidebarDropIndicator(tabId: second, edge: .top)
-            )
-        )
-    }
-
-    func testWorkspaceDropAfterLastRowFallsThroughToEmptyAreaDropTarget() {
-        let first = UUID()
-        let second = UUID()
-        let targets = workspaceDropTargets([first, second])
-
-        XCTAssertNil(
-            SidebarDropPlanner.workspaceAction(
-                for: CGPoint(x: 12, y: 92),
-                targets: targets
-            )
-        )
-    }
-
-    func testWorkspaceDropKeepsNewWorkspaceAfterPinnedRows() {
-        let pinnedA = UUID()
-        let pinnedB = UUID()
-        let unpinned = UUID()
-        let targets = workspaceDropTargets([pinnedA, pinnedB, unpinned], pinnedIds: [pinnedA, pinnedB])
-
-        let action = SidebarDropPlanner.workspaceAction(
-            for: CGPoint(x: 12, y: 2),
-            targets: targets
-        )
-
-        XCTAssertEqual(
-            action,
-            .newWorkspace(
-                insertionIndex: 2,
-                indicator: SidebarDropIndicator(tabId: unpinned, edge: .top)
-            )
-        )
-    }
-
-    private func workspaceDropTargets(
-        _ ids: [UUID],
-        pinnedIds: Set<UUID> = []
-    ) -> [SidebarDropPlanner.WorkspaceDropTarget] {
-        ids.enumerated().map { index, id in
-            SidebarDropPlanner.WorkspaceDropTarget(
-                workspaceId: id,
-                isPinned: pinnedIds.contains(id),
-                frame: CGRect(x: 0, y: CGFloat(index * 40), width: 180, height: 32)
-            )
-        }
-    }
 }
